add support for libsecp256k1
