--- buidl/answers.py
@@ -1,13 +1,0 @@
-"""
-#code
->>> import wallet
-
-#endcode
-"""
-
-from unittest import TestCase
-
-
-class SessionTest(TestCase):
-    def test_apply(self):
-        pass
